get experiment1
